+++ bundle.mjs
@@ -21,6 +21,7 @@ const bundle = async () => {
         },
       }),
     ],
+    external: ['commander', 'prompts'],
     tsconfig: './tsconfig.web.json',
     outfile: './bundles/web.bundle.min.js',
   })